Calculate finish dates, time intervals and clock when the activeTimeColumn is set
