+++ lib/Map/TableColumn.js
@@ -1,20 +1,14 @@
 /*global require*/
 "use strict";
 
-var ClockRange = require('terriajs-cesium/Source/Core/ClockRange');
-var ClockStep = require('terriajs-cesium/Source/Core/ClockStep');
-var DataSourceClock = require('terriajs-cesium/Source/DataSources/DataSourceClock');
 var defaultValue = require('terriajs-cesium/Source/Core/defaultValue');
 var defined = require('terriajs-cesium/Source/Core/defined');
 var defineProperties = require('terriajs-cesium/Source/Core/defineProperties');
 var deprecationWarning = require('terriajs-cesium/Source/Core/deprecationWarning');
 var destroyObject = require('terriajs-cesium/Source/Core/destroyObject');
 var DeveloperError = require('terriajs-cesium/Source/Core/DeveloperError');
-var Iso8601 = require('terriajs-cesium/Source/Core/Iso8601');
 var JulianDate = require('terriajs-cesium/Source/Core/JulianDate');
 var knockout = require('terriajs-cesium/Source/ThirdParty/knockout');
-var TimeInterval = require('terriajs-cesium/Source/Core/TimeInterval');
-var TimeIntervalCollection = require('terriajs-cesium/Source/Core/TimeIntervalCollection');
 
 var formatNumberForLocale = require('../Core/formatNumberForLocale');
 var getUniqueValues = require('../Core/getUniqueValues');
@@ -37,7 +31,6 @@ var subtypeHintSet = [
     { hint: /^(.*[_ ])?(year)/i, type: VarSubType.YEAR }
 ];
 
-var defaultFinalDurationSeconds = 3600 * 24 - 1; // one day less a second, if there is only one date.
 var defaultReplaceWithNullValues = ['na', 'NA', '-'];
 var defaultReplaceWithZeroValues = [];
 
@@ -68,7 +61,6 @@ var defaultReplaceWithZeroValues = [];
 * @param {String} [options.id] Provided so that columns can be renamed; their original name is stored as the id.
 * @param {String} [options.format] A format string for this column. For numbers, this is passed as options to toLocaleString.
 * @param {String} [options.units] The units of this column, if known. Not currently used internally by TableStructure or TableColumn.
-* @param {Boolean} [options.noFinishDates] If true, then finishDates, timeIntervals and clock are not calculated (even if this is a time column).
 * @param {String} [options.color] The string description of the color of this variable, if any.
 */
 var TableColumn = function(name, values, options) {
@@ -124,16 +116,19 @@ var TableColumn = function(name, values, options) {
     this.julianDates = undefined;
     /**
      * this.finishJulianDates is an Array of JulianDates listing the next different date in the values array, less 1 second.
+     * This is populated by TableStructure, since it may depend on other columns.
      * Only if type === VarType.TIME.
      */
     this.finishJulianDates = undefined;
     /**
      * A TimeInterval Array giving when each row applies.
+     * This is populated by TableStructure, since it may depend on other columns.
      * Only if type === VarType.TIME.
      */
     this._timeIntervals = undefined;
     /**
      * A DataSourceClock whose start and stop times correspond to the first and last visible row.
+     * This is populated by TableStructure, since it may depend on other columns.
      * Only if type === VarType.TIME.
      */
     this._clock = undefined;
@@ -147,12 +142,6 @@ var TableColumn = function(name, values, options) {
             this._type = VarType.SCALAR;
         } else {
             this._subtype = jsDatesAndJulianDates.subtype;
-            // Calculate default end dates and timeIntervals, and define a clock.
-            if (!this.options.noFinishDates) {
-                this.finishJulianDates = calculateFinishDates(this.julianDates);
-                this._timeIntervals = calculateTimeIntervals(this);
-                this._clock = createClock(this);
-            }
         }
     }
 
@@ -490,7 +479,6 @@ TableColumn.prototype.getFullOptions = function() {
         type: this._type,
         subtype: this._subtype,
         displayDuration: this.displayDuration,
-        noFinishDates: this.options.noFinishDates,
         displayVariableTypes: this._displayVariableTypes,
         color: this.color
     };
@@ -594,133 +582,6 @@ function convertToDates(tableColumn) {
     // Do this by replacing both sorts with null.
     var goodValues = replaceValues(tableColumn._values, [], tableColumn._replaceWithNullValues.concat(tableColumn._replaceWithZeroValues));
     return TableColumn.convertToDates(goodValues, tableColumn.subtype);
-}
-
-// Returns indices such that sortedUniqueDates[inverseIndices[k]] = originalDates[k].
-// Eg. var data = ['c', 'a', 'b', 'd'];
-//     var sortedData = data.slice().sort();
-//     var inverseIndices = data.map(function(datum) { return sortedData.indexOf(datum); });
-//     expect(inverseIndices).toEqual([2, 0, 1, 3]);
-// However this works by converting the dates to strings first.
-function calculateInverseIndicies(originalDates, sortedUniqueDates) {
-    var originalDateStrings = originalDates.map(function(date) { return date && JulianDate.toIso8601(date); });
-    var sortedUniqueDateStrings = sortedUniqueDates.map(function(date) { return date && JulianDate.toIso8601(date); });
-    return originalDateStrings.map(function(s) { return sortedUniqueDateStrings.indexOf(s); });
-}
-
-function calculateSortedUniqueJulianDates(originalDates) {
-    var sortedUniqueJulianDates = originalDates.filter(function(d) { return defined(d); });
-    sortedUniqueJulianDates.sort(JulianDate.compare);
-    sortedUniqueJulianDates = sortedUniqueJulianDates.filter(function(element, index, array) {
-        return (index === 0) || (!JulianDate.equals(array[index - 1], element));
-    });
-    return sortedUniqueJulianDates;
-}
-
-// For each date, find the next different date (minus 1 second). Return an array of these finish dates.
-// For the final date, use the average spacing of the unique dates as the final duration.
-// (If there is only one date, use a default value.)
-function calculateFinishDates(originalJulianDates) {
-    // First calculate a set of unique, sorted dates.
-    var sortedUniqueJulianDates = calculateSortedUniqueJulianDates(originalJulianDates);
-    // indices[k] has the property that originalJulianDates[indices[k]] = sortedUniqueJulianDates[k].
-    var indices = calculateInverseIndicies(originalJulianDates, sortedUniqueJulianDates);
-    // Calculate end dates corresponding to each revised date (which are start dates).
-    // Typically just shave a second off the next start date, unless the difference is < 20 seconds,
-    // in which case shave off 5% of the difference.
-    var endDates = sortedUniqueJulianDates.slice(1).map(function(rawEndDate, index) {
-        var secondsDifference = JulianDate.secondsDifference(rawEndDate, sortedUniqueJulianDates[index]);
-        if (secondsDifference < 20) {
-            return JulianDate.addSeconds(sortedUniqueJulianDates[index], secondsDifference * 0.95, new JulianDate());
-        } else {
-            return JulianDate.addSeconds(rawEndDate, -1, new JulianDate());
-        }
-    });
-    // For the final end date, use the average spacing of the unique dates.
-    // If there is only one date, use defaultFinalDurationSeconds.
-    var finalDurationSeconds = defaultFinalDurationSeconds;
-    var n = sortedUniqueJulianDates.length;
-    if (n > 1) {
-        finalDurationSeconds = JulianDate.secondsDifference(sortedUniqueJulianDates[n - 1], sortedUniqueJulianDates[0]) / (n - 1);
-    }
-    endDates.push(JulianDate.addSeconds(sortedUniqueJulianDates[n - 1], finalDurationSeconds, new JulianDate()));
-
-    var result = indices.map(function(sortedIndex) {
-        return endDates[sortedIndex];
-    });
-    return result;
-    // return originalJulianDates.map(function(startDate, index) {
-    //     if (defined(startDate)) {
-    //         for (var i = 0; i < sortedUniqueJulianDates.length; i++) {
-    //             if (JulianDate.greaterThan(endDates[i], startDate)) {
-    //                 return endDates[i];
-    //             }
-    //         }
-    //     }
-    // });
-}
-
-var endScratch = new JulianDate();
-/**
- * Calculate and return the availability interval for the index'th entry in timeColumn.
- * If the entry has no valid time, returns undefined.
- * @private
- * @param  {TableColumn} timeColumn The time column that applies to this data.
- * @param  {Integer} index The index into the time column.
- * @return {TimeInterval} The time interval over which this entry is visible.
- */
-function calculateAvailability(timeColumn, index) {
-    var startJulianDate = timeColumn.julianDates[index];
-    if (defined(startJulianDate)) {
-        var finishJulianDate;
-        if (!defined(timeColumn.displayDuration)) {
-            finishJulianDate = timeColumn.finishJulianDates[index];
-        } else {
-            finishJulianDate = JulianDate.addMinutes(timeColumn.julianDates[index], timeColumn.displayDuration, endScratch);
-        }
-        return new TimeInterval({start: timeColumn.julianDates[index], stop: finishJulianDate});
-    }
-}
-
-/**
- * Calculates and returns TimeInterval array, whose elements say when to display each row.
- * @private
- */
-function calculateTimeIntervals(timeColumn) {
-    return timeColumn.values.map(function(value, index) {
-        return calculateAvailability(timeColumn, index);
-    });
-}
-
-/**
- * Returns a DataSourceClock out of this column. Only call if this is a time column.
- * @private
- */
-function createClock(timeColumn) {
-    var availabilityCollection = new TimeIntervalCollection();
-    timeColumn._timeIntervals
-        .filter(function(availability) { return defined(availability && availability.start); })
-        .forEach(function(availability) {
-            availabilityCollection.addInterval(availability);
-        });
-    if (!defined(timeColumn._clock)) {
-        if (!availabilityCollection.start.equals(Iso8601.MINIMUM_VALUE)) {
-            var startTime = availabilityCollection.start;
-            var stopTime = availabilityCollection.stop;
-            var totalSeconds = JulianDate.secondsDifference(stopTime, startTime);
-            var multiplier = Math.round(totalSeconds / 120.0);
-
-            var clock = new DataSourceClock();
-            clock.startTime = JulianDate.clone(startTime);
-            clock.stopTime = JulianDate.clone(stopTime);
-            clock.clockRange = ClockRange.LOOP_STOP;
-            clock.multiplier = multiplier;
-            clock.currentTime = JulianDate.clone(startTime);
-            clock.clockStep = ClockStep.SYSTEM_CLOCK_MULTIPLIER;
-            return clock;
-        }
-    }
-    return timeColumn._clock;
 }
 
 // Returns true if all non-blank values could be html tags.
