+++ lib/Map/TableStructure.js
@@ -3,13 +3,19 @@
 
 var dateFormat = require('dateformat');
 
+var ClockRange = require('terriajs-cesium/Source/Core/ClockRange');
+var ClockStep = require('terriajs-cesium/Source/Core/ClockStep');
+var DataSourceClock = require('terriajs-cesium/Source/DataSources/DataSourceClock');
 var defaultValue = require('terriajs-cesium/Source/Core/defaultValue');
 var defined = require('terriajs-cesium/Source/Core/defined');
 var defineProperties = require('terriajs-cesium/Source/Core/defineProperties');
 var destroyObject = require('terriajs-cesium/Source/Core/destroyObject');
 var DeveloperError = require('terriajs-cesium/Source/Core/DeveloperError');
+var Iso8601 = require('terriajs-cesium/Source/Core/Iso8601');
 var JulianDate = require('terriajs-cesium/Source/Core/JulianDate');
 var knockout = require('terriajs-cesium/Source/ThirdParty/knockout');
+var TimeInterval = require('terriajs-cesium/Source/Core/TimeInterval');
+var TimeIntervalCollection = require('terriajs-cesium/Source/Core/TimeIntervalCollection');
 
 var csv = require('../ThirdParty/csv');
 var DisplayVariablesConcept = require('../Map/DisplayVariablesConcept');
@@ -18,6 +24,7 @@ var TableColumn = require('./TableColumn');
 var VarType = require('../Map/VarType');
 
 var defaultDisplayVariableTypes = [VarType.ENUM, VarType.SCALAR, VarType.ALT];
+var defaultFinalDurationSeconds = 3600 * 24 - 1; // one day less a second, if there is only one date.
 
 /**
  * TableStructure provides an abstraction of a data table, ie. a structure with rows and columns.
@@ -64,7 +71,7 @@ var TableStructure = function(name, options) {
      * Gets or sets the active time column name, id or index.
      * @type {String|Number|undefined}
      */
-    this.activeTimeColumnNameIdOrIndex = undefined;
+    this._activeTimeColumnNameIdOrIndex = undefined;
 
     // Track sourceFeature as it is shown on the NowViewing panel.
     // Track items so that charts can update live. (Already done by DisplayVariableConcept.)
@@ -129,14 +136,37 @@ defineProperties(TableStructure.prototype, {
     },
 
     /**
+     * Gets or sets the type of this column.
+     * @memberOf TableColumn.prototype
+     * @type {VarType}
+     */
+    activeTimeColumnNameIdOrIndex: {
+        get: function() {
+            return this._activeTimeColumnNameIdOrIndex;
+        },
+        set: function(nameIdOrIndex) {
+            this._activeTimeColumnNameIdOrIndex = nameIdOrIndex;
+            // Sort by the active time column, if available (so charts and derived charts don't double-back on themselves).
+            this.sortBy(this.activeTimeColumn);
+            var activeTimeColumn = this.activeTimeColumn;  // sortBy changes all the columns, so this.activeTimeColumn is now different.
+            if (defined(activeTimeColumn) && (!defined(activeTimeColumn.finishJulianDates))) {
+                // Calculate default end dates and timeIntervals, and define a clock on the active time column.
+                activeTimeColumn.finishJulianDates = calculateFinishDates(this, activeTimeColumn);
+                activeTimeColumn._timeIntervals = calculateTimeIntervals(activeTimeColumn);
+                activeTimeColumn._clock = createClock(activeTimeColumn);
+            }
+        }
+    },
+
+    /**
      * Gets the active time column for this structure.
      * @memberOf TableStructure.prototype
      * @type {TableColumn}
      */
     activeTimeColumn: {
         get: function() {
-            if (defined(this.activeTimeColumnNameIdOrIndex)) {
-                return this.getColumnWithNameIdOrIndex(this.activeTimeColumnNameIdOrIndex);
+            if (defined(this._activeTimeColumnNameIdOrIndex)) {
+                return this.getColumnWithNameIdOrIndex(this._activeTimeColumnNameIdOrIndex);
             }
         }
     },
@@ -148,13 +178,27 @@ defineProperties(TableStructure.prototype, {
      */
     timeIntervals: {
         get: function() {
-            if (!defined(this.activeTimeColumnNameIdOrIndex)) {
+            var activeTimeColumn = this.activeTimeColumn;
+            if (!defined(this.activeTimeColumn)) {
                 return;
             }
-            if (!defined(this._timeIntervals)) {
-                this._timeIntervals = calculateTimeIntervals(this);
+            return activeTimeColumn._timeIntervals;
+        }
+    },
+
+    /**
+     * Returns a clock whose start and stop times correspond to the first and last visible row.
+     * Only defined if type == VarType.TIME.
+     * @memberOf TableColumn.prototype
+     * @type {DataSourceClock}
+     */
+    clock: {
+        get: function() {
+            var activeTimeColumn = this.activeTimeColumn;
+            if (!defined(this.activeTimeColumn)) {
+                return;
             }
-            return this._timeIntervals;
+            return activeTimeColumn._clock;
         }
     }
 
@@ -292,25 +336,36 @@ TableStructure.prototype.getActiveColumns = function() {
     return this.columns.filter(function(column) { return column.isActive; });
 };
 
-/**
- * A Table might have dates 2015-03-01, 2015-04-01
- * @param  {[type]} tableStructure [description]
- * @return {[type]}                [description]
- * @private
- */
-function calculateNaturalPeriodLength(tableStructure) {
 
+// Returns indices such that sortedUniqueDates[inverseIndices[k]] = originalDates[k].
+// Eg. var data = ['c', 'a', 'b', 'd'];
+//     var sortedData = data.slice().sort();
+//     var inverseIndices = data.map(function(datum) { return sortedData.indexOf(datum); });
+//     expect(inverseIndices).toEqual([2, 0, 1, 3]);
+// However this works by converting the dates to strings first.
+function calculateInverseIndicies(originalDates, sortedUniqueDates) {
+    var originalDateStrings = originalDates.map(function(date) { return date && JulianDate.toIso8601(date); });
+    var sortedUniqueDateStrings = sortedUniqueDates.map(function(date) { return date && JulianDate.toIso8601(date); });
+    return originalDateStrings.map(function(s) { return sortedUniqueDateStrings.indexOf(s); });
 }
 
+function calculateSortedUniqueJulianDates(originalDates) {
+    var sortedUniqueJulianDates = originalDates.filter(function(d) { return defined(d); });
+    sortedUniqueJulianDates.sort(JulianDate.compare);
+    sortedUniqueJulianDates = sortedUniqueJulianDates.filter(function(element, index, array) {
+        return (index === 0) || (!JulianDate.equals(array[index - 1], element));
+    });
+    return sortedUniqueJulianDates;
+}
 
 // For each row, find the next different date (minus 1 second).
-// Restrict to only those rows with this value of the idColumnNames, if present.
+// TODO: Restrict to only those rows with this value of the idColumnNames, if present.
 // Return an array of these finish dates, one per row.
 // Assume the rows are already sorted by date.
 // For the final date, use the average spacing of the unique dates as the final duration.
 // (If there is only one date, use a default value.)
-function calculateFinishDates(tableStructure) {
-    var originalJulianDates = tableStructure.activeTimeColumn.julianDates;
+function calculateFinishDates(tableStructure, timeColumn) {
+    var originalJulianDates = timeColumn.julianDates;
     // First calculate a set of unique, sorted dates.
     var sortedUniqueJulianDates = calculateSortedUniqueJulianDates(originalJulianDates);
     // indices[k] has the property that originalJulianDates[indices[k]] = sortedUniqueJulianDates[k].
@@ -371,6 +426,37 @@ function calculateTimeIntervals(timeColumn) {
     return timeColumn.values.map(function(value, index) {
         return calculateAvailability(timeColumn, index);
     });
+}
+
+/**
+ * Returns a DataSourceClock out of this column. Only call if this is a time column.
+ * @private
+ */
+function createClock(timeColumn) {
+    var availabilityCollection = new TimeIntervalCollection();
+    timeColumn._timeIntervals
+        .filter(function(availability) { return defined(availability && availability.start); })
+        .forEach(function(availability) {
+            availabilityCollection.addInterval(availability);
+        });
+    if (!defined(timeColumn._clock)) {
+        if (!availabilityCollection.start.equals(Iso8601.MINIMUM_VALUE)) {
+            var startTime = availabilityCollection.start;
+            var stopTime = availabilityCollection.stop;
+            var totalSeconds = JulianDate.secondsDifference(stopTime, startTime);
+            var multiplier = Math.round(totalSeconds / 120.0);
+
+            var clock = new DataSourceClock();
+            clock.startTime = JulianDate.clone(startTime);
+            clock.stopTime = JulianDate.clone(stopTime);
+            clock.clockRange = ClockRange.LOOP_STOP;
+            clock.multiplier = multiplier;
+            clock.currentTime = JulianDate.clone(startTime);
+            clock.clockStep = ClockStep.SYSTEM_CLOCK_MULTIPLIER;
+            return clock;
+        }
+    }
+    return timeColumn._clock;
 }
