+++ lib/Models/TableCatalogItem.js
@@ -469,14 +469,6 @@ TableCatalogItem.prototype.initializeFromTableStructure = function(tableStructur
     var tableStyle = item._tableStyle;
 
     tableStructure.setActiveTimeColumn(tableStyle.timeColumn);
-    // Sort by the active time column, if available (so charts and derived charts don't double-back on themselves).
-    if (defined(tableStructure.activeTimeColumn)) {
-        // Note this means we wind up calculating the availabilities in TableColumn twice, which is unnecessarily slow.
-        // Rather than adding complexity to CsvCatalogItem to turn off that calculation initially, and then add it back here,
-        // it may be better to refactor the availability calculation out of TableColumn, as part of the fix for
-        // https://github.com/TerriaJS/terriajs/issues/2018.
-        tableStructure.sortBy(tableStructure.activeTimeColumn);
-    }
     item._tableStructure = tableStructure;
 
     function makeChartable() {
@@ -587,10 +579,6 @@ function createDataSourceForLatLong(item, tableStructure) {
 
 TableCatalogItem.prototype.setChartable = function() {
     var tableStructure = this._tableStructure;
-    // Any derived calculations from this can ignore the need for julianFinishDates and time intervals.
-    tableStructure.columnsByType[VarType.TIME].forEach(function(column) {
-        column.options.noFinishDates = true;
-    });
     tableStructure.allowMultiple = true;
     this.isMappable = false;
     tableStructure.getColorCallback = this.getNextColor.bind(this);
